backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
django==4.2.23
django-filter==24.3
django-isbn-field==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
importlib-resources==6.4.5
inflection==0.5.1
isbnlib==3.10.14
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==25.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-stdnum==2.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
zipp==3.20.2
